Restructure  loan API
