+++ src/core/ddsc/CMakeLists.txt
@@ -89,7 +89,7 @@ prepend(hdrs_public_ddsc "$<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/include>$<
   ddsc/dds_internal_api.h
   ddsc/dds_opcodes.h
   ddsc/dds_data_allocator.h
-  ddsc/dds_loan.h)
+  ddsc/dds_loan_api.h)
 
 if (DDS_HAS_SHM)
   list(APPEND hdrs_private_ddsc "${CMAKE_CURRENT_LIST_DIR}/src/shm__monitor.h")
