+++ src/core/ddsi/CMakeLists.txt
@@ -177,8 +177,8 @@ if(ENABLE_SECURITY)
   list(APPEND hdrs_private_ddsi ddsi_security_msg.h ddsi_security_exchange.h)
 endif()
 if(iceoryx_binding_c_FOUND)
-  list(APPEND srcs_ddsi shm_transport.c) 
-  list(APPEND hdrs_private_ddsi shm_transport.h)
+  list(APPEND srcs_ddsi ddsi_shm_transport.c) 
+  list(APPEND hdrs_private_ddsi ddsi_shm_transport.h)
 endif()
 
 prepend(hdrs_private_ddsi "${CMAKE_CURRENT_LIST_DIR}/include/dds/ddsi/" ${hdrs_private_ddsi})
