+++ src/core/ddsi/include/dds/ddsi/ddsi_shm_transport.h
@@ -12,19 +12,20 @@
 #ifndef DDS_SHM__TRANSPORT_H
 #define DDS_SHM__TRANSPORT_H
 
+#include "dds/export.h"
+#include "dds/features.h"
+
 #include "dds/ddsi/ddsi_config.h"
 #include "dds/ddsi/ddsi_keyhash.h"
 #include "dds/ddsi/ddsi_tran.h"
 #include "dds/ddsi/q_protocol.h" /* for, e.g., SubmessageKind_t */
 #include "dds/ddsrt/sync.h"
-#include "dds/export.h"
-#include "dds/features.h"
 
 #include "iceoryx_binding_c/chunk.h"
 #include "iceoryx_binding_c/publisher.h"
 #include "iceoryx_binding_c/subscriber.h"
 
-#if defined (__cplusplus)
+#if defined(__cplusplus)
 extern "C" {
 #endif
 
@@ -35,13 +36,13 @@ typedef enum {
 } iox_shm_data_state_t;
 
 struct iceoryx_header {
-   struct ddsi_guid guid;
-   dds_time_t tstamp;
-   uint32_t statusinfo;
-   uint32_t data_size;
-   unsigned char data_kind;
-   ddsi_keyhash_t keyhash;
-   iox_shm_data_state_t shm_data_state;
+  struct ddsi_guid guid;
+  dds_time_t tstamp;
+  uint32_t statusinfo;
+  uint32_t data_size;
+  unsigned char data_kind;
+  ddsi_keyhash_t keyhash;
+  iox_shm_data_state_t shm_data_state;
 };
 
 typedef struct iceoryx_header iceoryx_header_t;
@@ -50,17 +51,20 @@ struct dds_reader;
 struct shm_monitor;
 
 typedef struct {
-    iox_sub_storage_t storage;
-    //we use a mutex per subscriber to handle concurrent take and release of the data
-    ddsrt_mutex_t mutex;
-    struct shm_monitor* monitor;
-    struct dds_reader* parent_reader;
+  iox_sub_storage_t storage;
+  // we use a mutex per subscriber to handle concurrent take and release of the
+  // data
+  ddsrt_mutex_t mutex;
+  struct shm_monitor *monitor;
+  struct dds_reader *parent_reader;
 } iox_sub_storage_extension_t;
 
-//lock and unlock for individual subscribers
-DDS_EXPORT void iox_sub_storage_extension_init(iox_sub_storage_extension_t* storage);
+// lock and unlock for individual subscribers
+DDS_EXPORT void
+iox_sub_storage_extension_init(iox_sub_storage_extension_t *storage);
 
-DDS_EXPORT void iox_sub_storage_extension_fini(iox_sub_storage_extension_t* storage);
+DDS_EXPORT void
+iox_sub_storage_extension_fini(iox_sub_storage_extension_t *storage);
 
 DDS_EXPORT void shm_lock_iox_sub(iox_sub_t sub);
 
@@ -79,8 +83,8 @@ DDS_EXPORT void shm_set_data_state(void *iox_chunk,
 
 DDS_EXPORT iox_shm_data_state_t shm_get_data_state(void *iox_chunk);
 
-#if defined (__cplusplus)
+#if defined(__cplusplus)
 }
 #endif
 
-#endif //DDS_SHM__TRANSPORT_H
+#endif // DDS_SHM__TRANSPORT_H
